requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-redis==6.0.0
gunicorn==21.2.0
packaging==25.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==7.1.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.7.0
